Use int64 IDs for jobs

+++ internal/storage/redis/store.go
@@ -11,7 +11,7 @@ type Store struct {
 }
 
 // Get returns a Job object by its ID.
-func (s *Store) Get(ID int) (*storage.Job, error) {
+func (s *Store) Get(ID int64) (*storage.Job, error) {
 
 	return nil, nil
 }
@@ -23,7 +23,7 @@ func (s *Store) Put(job storage.Job) error {
 }
 
 // Delete removes a Job from the store.
-func (s *Store) Delete(ID int) error {
+func (s *Store) Delete(ID int64) error {
 
 	return nil
 }
